Functionality of Database implemented (DAO)
Methods in DCServer are implemented to be used with DTOs later on. Creating a new object of the same class will be deprecated and changed by DTOtoCustomerData() method as this functions are called from the client.

Changed DAOs to work with CustomerData as Customer class is now unused.

Fixed package path minor problem.